#279 move drag and drop to actions
