+++ lib/handlers/inputHandler.js
@@ -167,10 +167,15 @@ const _insertText = async (char) => {
   await input.insertText({ text: char });
 };
 
+const dispatchMouseEvent = async (options) => {
+  await input.dispatchMouseEvent(options);
+};
+
 module.exports = {
   up: _up,
   down: _down,
   mouse_move: _mouse_move,
   sendCharacter: _insertText,
   tap,
+  dispatchMouseEvent,
 };
