+++ lib/taiko.js
@@ -26,6 +26,7 @@ const targetHandler = require('./handlers/targetHandler');
 const runtimeHandler = require('./handlers/runtimeHandler');
 const browserHandler = require('./handlers/browserHandler');
 const emulationHandler = require('./handlers/emulationHandler');
+const { description } = require('./actions/pageActionChecks');
 const { logEvent } = require('./logger');
 const {
   match,
@@ -1262,66 +1263,8 @@ module.exports.rightClick = async (selector, options = {}, ...args) => {
  */
 module.exports.dragAndDrop = async (source, destination) => {
   validate();
-  let sourceElem = await findFirstElement(source);
-  let dest =
-    isSelector(destination) || isString(destination) || isElement(destination)
-      ? await findFirstElement(destination)
-      : destination;
-  let options = setClickOptions({});
-  await doActionAwaitingNavigation(options, async () => {
-    if (defaultConfig.headful) {
-      await highlightElement(sourceElem);
-      if (isElement(dest)) {
-        await highlightElement(dest);
-      }
-    }
-    await dragAndDrop(options, sourceElem, dest);
-  });
-  const desc = isElement(dest)
-    ? `Dragged and dropped ${description(sourceElem, true)} to ${description(dest, true)}}`
-    : `Dragged and dropped ${description(sourceElem, true)} at ${JSON.stringify(destination)}`;
+  const desc = require('./actions/dragAndDrop').dragAndDrop(source, destination);
   descEvent.emit('success', desc);
-};
-
-const dragAndDrop = async (options, sourceElem, dest) => {
-  let sourcePosition = await domHandler.boundingBoxCenter(sourceElem.get());
-  await scrollTo(sourceElem);
-  options.x = sourcePosition.x;
-  options.y = sourcePosition.y;
-  options.type = 'mouseMoved';
-  await input.dispatchMouseEvent(options);
-  options.type = 'mousePressed';
-  await input.dispatchMouseEvent(options);
-  let destPosition = await calculateDestPosition(sourceElem.get(), dest);
-  await inputHandler.mouse_move(sourcePosition, destPosition);
-  options.x = destPosition.x;
-  options.y = destPosition.y;
-  options.type = 'mouseReleased';
-  await input.dispatchMouseEvent(options);
-};
-
-const calculateDestPosition = async (sourceElemobjectId, dest) => {
-  if (isElement(dest)) {
-    await scrollTo(dest);
-    return await domHandler.boundingBoxCenter(dest.get());
-  }
-  const destPosition = await domHandler.calculateNewCenter(sourceElemobjectId, dest);
-  const newBoundary = destPosition.newBoundary;
-  if (defaultConfig.headful) {
-    await overlayHandler.highlightQuad([
-      newBoundary.right,
-      newBoundary.top,
-      newBoundary.right,
-      newBoundary.bottom,
-      newBoundary.left,
-      newBoundary.bottom,
-      newBoundary.left,
-      newBoundary.top,
-    ]);
-    await waitFor(1000);
-    await overlayHandler.hideHighlight();
-  }
-  return destPosition;
 };
 
 /**
@@ -1532,7 +1475,9 @@ module.exports.clear = async (selector, options = {}) => {
   await doActionAwaitingNavigation(options, async () => {
     await _clear(activeElement.objectId);
     if (defaultConfig.headful) {
-      await highlightElement(new Element(activeElement.objectId, '', runtimeHandler));
+      await highlightElement(
+        new Element(activeElement.objectId, '', runtimeHandler),
+      );
     }
   });
   descEvent.emit('success', desc);
@@ -3209,18 +3154,6 @@ module.exports.getConfig = getConfig;
  * <a href="#press">press</a> and <a href="#evaluate">evaluate</a>.
  */
 module.exports.setConfig = setConfig;
-
-const description = (selector, lowerCase = false) => {
-  const d = (() => {
-    if (isString(selector)) {
-      return `Element matching text "${selector}"`;
-    } else if (isSelector(selector) || isElement(selector)) {
-      return selector.description;
-    }
-    return '';
-  })();
-  return lowerCase ? d.charAt(0).toLowerCase() + d.slice(1) : d;
-};
 
 const waitForMouseActions = async (options) => {
   await doActionAwaitingNavigation(options, async () => {
